Rename side-menu-content.tsx -> SideMenuContent
