+++ .gitignore
@@ -763,9 +763,9 @@ packages/app-mobile/components/screens/status.js
 packages/app-mobile/components/screens/tags.js
 packages/app-mobile/components/sidebar/SideMenu.js
 packages/app-mobile/components/sidebar/SideMenuButton.js
+packages/app-mobile/components/sidebar/SideMenuContent.js
 packages/app-mobile/components/sidebar/SideMenuContentNote.js
 packages/app-mobile/components/sidebar/SyncButton.js
-packages/app-mobile/components/sidebar/side-menu-content.js
 packages/app-mobile/components/testing/TestProviderStack.js
 packages/app-mobile/components/voiceTyping/VoiceTypingDialog.js
 packages/app-mobile/gulpfile.js
